pyproject.toml
[project]
name = "ember-ai"
version = "0.1.0"
description = "A compositional framework for building Compound AI Systems with JAX-style transforms."
authors = [
    {name = "Jared Quincy Davis", email = "jared@pyember.org"}
]
maintainers = [
    {name = "Ember Team", email = "team@pyember.org"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.14"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Operating System :: OS Independent",
]
keywords = [
    "AI",
    "LLM",
    "Compound AI",
    "Machine Learning",
    "Orchestration",
    "AI Framework",
    "LLM Orchestration",
    "JAX",
    "JIT",
    "Operators",
]

dependencies = [
    # Core framework
    "equinox>=0.11.0",
    "jax>=0.4.0",
    "optax>=0.1.7",
    "numpy>=2.1.0",
    "pandas>=2.2.3,<2.4.0",
    "pydantic>=2.7.4",
    "pydantic-core>=2.18.4",
    "pydantic-settings>=2.3.0",
    "PyYAML>=6.0.1",
    "typing_extensions>=4.12.2",

    # LLM providers
    "openai>=2.5.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.3",
    "google-ai-generativelanguage>=0.6.6",

    # HTTP and async
    "aiohttp>=3.9.5",
    "aiosignal>=1.3.1",
    "httpx==0.25.2",
    "httpcore==1.0.2",
    "h11==0.14.0",
    "requests>=2.32.2",

    # Data processing
    "datasets>=2.20.0",
    "torch>=2.3.0",
    "transformers>=4.57.0",
    "sentence-transformers>=3.0.0",
    "huggingface-hub>=0.35.3,<1.0.0",

    # Utilities
    "rich>=13.0.0",
    "tqdm>=4.67.1",
    "tenacity>=9.0.0",
    "cachetools>=5.4.0",
    "dill>=0.3.8",
    "tiktoken>=0.7.0",
    "prettytable>=3.12.0",
    "beartype>=0.17.2",

    # gRPC (for Google)
    "grpcio>=1.71.0,<1.72.0",
    "grpcio-status>=1.71.0,<1.72.0",
    "proto-plus>=1.24.0,<2.0.0",
    "googleapis-common-protos>=1.63.2",
    "protobuf>=5.26.1,<6.0.0",

    # Compatibility
    "pyasn1==0.4.8",
    "pyasn1-modules==0.2.8",
    "annotated-types>=0.7.0",
    "anyio>=4.4.0",
    "attrs>=23.2.0",
    "certifi>=2024.6.2",
    "charset-normalizer>=3.3.2",
    "packaging>=24.1",
    "six>=1.16.0",
    "idna>=3.7",
    "urllib3>=1.26.19,<2.0.0",
    "sniffio>=1.3.1",
]

[project.optional-dependencies]
# Minimal installation with just OpenAI
minimal = ["openai>=2.5.0"]

# All providers
allproviders = [
    "openai>=2.5.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.3",
]

# Individual providers
openai = ["openai>=2.5.0"]
anthropic = ["anthropic>=0.40.0"]
google = [
    "google-generativeai>=0.8.3",
    "google-ai-generativelanguage>=0.6.6",
    "google-api-core>=2.19.1",
    "google-api-python-client>=2.139.0",
    "google-auth>=2.32.0",
    "google-auth-httplib2>=0.2.0",
    "googleapis-common-protos>=1.63.2",
]

# Data processing extras
data = [
    "datasets>=2.20.0",
    "scikit-learn>=1.6.0",
    "scipy>=1.13.1",
    "huggingface-hub>=0.26.5",
    "pyarrow>=16.1.0",
    "pyarrow-hotfix>=0.6",
]

# Visualization
viz = [
    "matplotlib>=3.9.1",
    "prettytable>=3.12.0",
]

# Development dependencies
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.8.0",
    "hypothesis>=6.99.0",
    "parameterized>=0.9.0",
    "black>=23.12.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "types-PyYAML>=6.0.12.20240808",
    "ruff>=0.1.6",
    "pre-commit>=3.5.0",
]

# Documentation
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "nbsphinx>=0.9.3",
    "myst-parser>=2.0.0",
    "jupyterlab>=4.0.6",
    "jupyter",
    "notebook",
]

# Full installation (all optional deps except dev/docs)
all = [
    "ember-ai[allproviders,data,viz]",
]

[project.urls]
Repository = "https://github.com/pyember/ember"
Documentation = "https://docs.pyember.org"
Homepage = "https://pyember.org"
"Bug Tracker" = "https://github.com/pyember/ember/issues"
Changelog = "https://github.com/pyember/ember/blob/main/CHANGELOG.md"

[project.scripts]
ember = "ember.cli.main:main"

[project.entry-points."ember.dataset_preppers"]
truthful_qa = "ember.core.utils.data.datasets_registry.truthful_qa:TruthfulQAPrepper"
short_answer = "ember.core.utils.data.datasets_registry.short_answer:ShortAnswerPrepper"
commonsense_qa = "ember.core.utils.data.datasets_registry.commonsense_qa:CommonsenseQAPrepper"
halueval = "ember.core.utils.data.datasets_registry.halueval:HaluEvalPrepper"
mmlu = "ember.core.utils.data.datasets_registry.mmlu:MMLUPrepper"

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--import-mode=importlib"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use external resources)",
    "slow: Slow running tests",
    "requires_api_key: Tests that require API keys",
    "asyncio: Async tests",
]
asyncio_mode = "strict"

[tool.coverage.run]
source = ["src/ember"]
omit = ["*/__init__.py", "*/test_*.py", "tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING",
    "pass",
    "\\.\\.\\."
]
show_missing = true

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
